pyproject.toml
[project]
name = "idealista-web-scraper"
version = "0.1.0"
description = "A web scraper for Idealista real estate listings in Portugal"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core
    "sqlalchemy>=2.0",
    "pydantic>=2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "typer>=0.9",
    # Scraping
    "beautifulsoup4>=4.14.2",
    "requests>=2.32.5",
    "httpx>=0.27",
    "lxml>=5.0",
    "playwright>=1.40",
    # Export
    "pandas>=2.0",
    "pyarrow>=15.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8",
    "ruff>=0.3",
    "pytest>=8.0",
    "pytest-cov>=4.0",
    # Type stubs
    "types-requests>=2.31",
    "types-beautifulsoup4>=4.12",
    "types-PyYAML>=6.0",
    "pandas-stubs>=2.0",
]

[project.scripts]
idealista-scraper = "idealista_scraper.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/idealista_scraper"]

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["idealista_scraper"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
